move HiGenCleaner plugin instantiation in plugins
